Implement API token authentication for upload requests
This commit refactors the `client` package to provide a single `Client` interface that is used when performing
uploads. The `client` package now has multiple `Client` implementations depending on the authentication method
desired in the agent configuration.

It includes the `OAuthClient`, `APITokenClient` and `UnauthenticatedClient` types which all implement `Client`
and use a respective authentication method/no authentication. The `APITokenClient` works in the same way as
the `OAuthClient` but uses an api token in the `Authorization` header instead of the oauth token.

The agent can now be configured using the `api-token` flag or the `API_TOKEN` environment variable.
